+++ src/main/java/roomescape/common/GlobalExceptionHandler.java
@@ -1,11 +1,11 @@
-package roomescape.reservation.common;
+package roomescape.common;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import roomescape.reservation.exception.ReservationNotFoundException;
-import roomescape.reservation.exception.ReservationTimeNotFoundException;
+import roomescape.reservationtime.exception.ReservationTimeNotFoundException;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
